# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internal API",
        "title": "Bigfood",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/auth/refresh-token": {
            "put": {
                "description": "Refresh user refresh and access tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh jwt token",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/refreshToken.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/refreshToken.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid refresh token format",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/cafe/user": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Edit cafe user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafe user"
                ],
                "summary": "Edit cafe user",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cafeUserEdit.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/actions.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "422": {
                        "description": "Owner role could not be set",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create cafe user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cafe user"
                ],
                "summary": "Create cafe user",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cafeUserCreate.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/actions.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "422": {
                        "description": "Cafe user already exist",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete cafe user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "cafe user"
                ],
                "summary": "Delete cafe user",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cafeUserDelete.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/table": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Edit table",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Edit table",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tableEdit.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create table",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Create table",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tableCreate.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/table.Table"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete table",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Delete table",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tableDelete.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            }
        },
        "/table/create-mass": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create N-quantity of tables",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "table"
                ],
                "summary": "Mass creation",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createMass.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/controller.TableListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Access Denied",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "actions.Response": {
            "type": "object",
            "properties": {
                "cafe-id": {
                    "type": "string",
                    "example": "uuid"
                },
                "comment": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "uuid"
                },
                "roles": {
                    "description": "todo: get list or roles from const",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "owner",
                        "admin",
                        "hostess"
                    ]
                },
                "user-id": {
                    "type": "string",
                    "example": "uuid"
                },
                "user.Name": {
                    "type": "string"
                }
            }
        },
...
        "cafeUserCreate.Message": {
            "type": "object",
            "required": [
                "cafe-id",
                "phone",
                "roles"
            ],
            "properties": {
                "cafe-id": {
                    "type": "string",
                    "example": "uuid"
                },
                "comment": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "example": "User phone"
                },
                "roles": {
                    "description": "todo: collection",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "owner",
                        "admin",
                        "hostess"
                    ]
                }
            }
        },
        "cafeUserDelete.Message": {
            "type": "object",
            "required": [
                "cafe-user-id"
            ],
            "properties": {
                "cafe-user-id": {
                    "type": "string",
                    "example": "uuid"
                }
            }
        },
        "cafeUserEdit.Message": {
            "type": "object",
            "required": [
                "cafe-user-id"
            ],
            "properties": {
                "cafe-user-id": {
                    "type": "string",
                    "example": "uuid"
                },
                "comment": {
                    "type": "string"
                },
                "roles": {
                    "description": "todo: bad array swagger and parse collection",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "controller.TableListResponse": {
            "type": "object",
            "properties": {
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/table.Table"
                    }
                }
            }
        },
...
        "createMass.Message": {
            "type": "object",
            "required": [
                "cafe-id",
                "quantity"
            ],
            "properties": {
                "cafe-id": {
                    "type": "string",
                    "example": "uuid"
                },
                "quantity": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "refreshToken.Message": {
            "type": "object",
            "required": [
                "refresh-token"
            ],
            "properties": {
                "refresh-token": {
                    "type": "string",
                    "example": "UUID"
                }
            }
        },
        "refreshToken.Response": {
            "type": "object",
            "properties": {
                "access-token": {
                    "type": "string"
                },
                "refresh-token": {
                    "type": "string",
                    "example": "UUID"
                },
                "user-id": {
                    "type": "string",
                    "example": "UUID"
                }
            }
        },
...
        "server.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error message"
                }
            }
        },
        "table.Table": {
            "type": "object",
            "properties": {
                "cafe-id": {
                    "type": "string",
                    "example": "uuid"
                },
                "comment": {
                    "type": "string",
                    "example": "comment"
                },
                "id": {
                    "type": "string",
                    "example": "uuid"
                },
                "seats": {
                    "type": "integer",
                    "example": 4
                },
                "title": {
                    "type": "string",
                    "example": "serial number"
                }
            }
        },
        "tableCreate.Message": {
            "type": "object",
            "required": [
                "cafe-id"
            ],
            "properties": {
                "cafe-id": {
                    "type": "string",
                    "example": "uuid"
                },
                "comment": {
                    "type": "string"
                },
                "seats": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "tableDelete.Message": {
            "type": "object",
            "required": [
                "table-id"
            ],
            "properties": {
                "table-id": {
                    "type": "string",
                    "example": "uuid"
                }
            }
        },
...
        "tableEdit.Message": {
            "type": "object",
            "required": [
                "table-id"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "seats": {
                    "type": "integer"
                },
                "table-id": {
                    "type": "string",
                    "example": "uuid"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}